Re-organize reqResp encoders code
