+++ packages/lodestar/src/network/reqresp/encodingStrategies/sszSnappy/snappyFrames/uncompress.ts
@@ -1,10 +1,9 @@
 import BufferList from "bl";
 import {uncompress} from "snappyjs";
-import {IDecompressor} from "../interface";
 
 const IDENTIFIER = Buffer.from([0x73, 0x4e, 0x61, 0x50, 0x70, 0x59]);
 
-export class SnappyFramesUncompress implements IDecompressor {
+export class SnappyFramesUncompress {
   private buffer = new BufferList();
 
   private state: IUncompressState = {
