+++ packages/lodestar/src/network/reqresp/interface.ts
@@ -1,8 +1,8 @@
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
-import {RequestId, RequestBody} from "@chainsafe/lodestar-types";
+import {RequestId, RequestBody, ResponseBody} from "@chainsafe/lodestar-types";
 import {ILogger} from "@chainsafe/lodestar-utils";
 import {ReqEventEmitter} from "..";
-import {Method, ReqRespEncoding} from "../../constants";
+import {Method, Methods, ReqRespEncoding, RpcResponseStatus} from "../../constants";
 import {IPeerMetadataStore, IRpcScoreTracker} from "../peers";
 
 export interface IReqEventEmitterClass {
@@ -23,3 +23,22 @@ export type ReqRespRequest<Body extends RequestBody | null = null> = {
   body: Body;
   encoding: ReqRespEncoding;
 };
+
+export interface IResponseChunk {
+  requestId: RequestId;
+  status: RpcResponseStatus;
+  body: ResponseBody;
+}
+
+export interface IValidatedRequestBody {
+  isValid: boolean;
+  // missing body if isValid=false
+  body?: RequestBody;
+}
+
+export type RequestOrResponseType = Exclude<
+  ReturnType<typeof Methods[Method]["responseSSZType"]> | ReturnType<typeof Methods[Method]["requestSSZType"]>,
+  null
+>;
+
+export type RequestOrResponseBody = ResponseBody | RequestBody;
